# openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "wish-api",
		"version": "0.1.0"
	},
...
	"paths": {
...
		"/auth/register": {
			"post": {
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"token": {
										"type": "string"
									},
									"response": {
										"type": "object",
										"properties": {},
										"additionalProperties": {}
									}
								},
								"additionalProperties": {}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Default Response"
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"type": "string"
									},
									"response": {}
								},
								"additionalProperties": {}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Default Response"
					}
				}
			}
		},
...
	}
}